[WIP] Bug 540499 - "Validation" context menu entry appears everywhere
The WTP plug-in org.eclipse.wst.validation.ui defines a context menu
entry "Validate" at the global context menu, in the additions section
(the section which contains e.g. "Run As" and "Debug As"). The
validation action, which backs up the context menu entry, then consists
of different WTP contributions.

The "Validation" menu entry is not always desired, e.g. when using
single features from WTP. Mainly, this is the case due to the central
positioning of the menu entry and its vague label.

This change adds a category for the "Validate" action. The category is
labeled "WebTools" and is located among the edit actions (among e.g.
"Source" and "Refactor"). With the category, its somewhat more clear
what is being validated.

Furthermore a system property
"org.eclipse.wst.validation.ui.disable.validation.context.menu" is
added. Setting this property to "true" will hide the "Validation"
context menu entry. The property can be used in products which take only
some parts of WTP and do not require the validation action. The property
can be set with a VM argument:

-Dorg.eclipse.wst.validation.ui.disable.validation.context.menu=true

Change-Id: I3fb400d312fd5ba04b3c0bd62274e888f4ce32d1
Signed-off-by: Simeon Andreev <simeon.danailov.andreev@gmail.com>